CoastCapitalPlatform/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0
requests>=2.31.0
anthropic>=0.40.0
mcp>=1.0.0
mysql-connector-python>=8.3.0
jinja2>=3.1.0
